Increase the size of struct pkgCache::Header::HeaderSz
Our compiler says that "unsigned short int" is not enough
to hold the value of sizeof(pkgCache::Header):

pkgcache.cc: In constructor 'pkgCache::Header::Header()':
pkgcache.cc:70:15: warning: conversion from 'long unsigned int' to 'short unsigned int' changes value from '262424' to '280' [-Woverflow]
   70 |    HeaderSz = sizeof(pkgCache::Header);
      |               ^~~~~~~~~~~~~~~~~~~~~~~~

This is an ABI change.

+++ apt-pkg/pkgcache.h
@@ -173,7 +173,7 @@ struct pkgCache::Header
    unsigned long OptionsHash;
 
    // Size of structure values
-   unsigned short HeaderSz;
+   unsigned int HeaderSz;
    unsigned short PackageSz;
    unsigned short PackageFileSz;
    unsigned short VersionSz;
